Algorithm

This section contains information about interface.jl

Interface

Boscia.BLMOStatus — Type

Enum encoding the status of the Bounded Linear Minimization Oracle.

Boscia.BoundedLinearMinimizationOracle — Type

BLMO

Supertype for the Bounded Linear Minimization Oracles

Boscia.add_bound_constraint! — Function

Add bound constraint.

Boscia.build_LMO_correct — Method

Check if the bounds were set correctly in build_LMO. Safety check only.

Boscia.build_global_bounds — Function

Read global bounds from the problem.

Boscia.check_feasibility — Method

Check if problem is bounded and feasible, i.e. no contradicting constraints.

Boscia.check_infeasible_vertex — Method

Deal with infeasible vertex if necessary, e.g. check what caused it etc.

Boscia.delete_bounds! — Function

Delete bounds.

Boscia.find_best_solution — Method

Find best solution from the solving process.

Boscia.free_model — Method

Free model data from previous solve (if necessary).

Boscia.get_BLMO_solve_data — Method

Get solve time, number of nodes and number of iterations, if applicable.

Boscia.get_bound — Function

Read bound value for c_idx.

Boscia.get_int_var — Function

Get the index of the integer variable the bound is working on.

Boscia.get_integer_variables — Function

Get list of integer variables.

Boscia.get_list_of_variables — Function

Get list of variables indices. If the problem has n variables, they are expected to contiguous and ordered from 1 to n.

Boscia.get_lower_bound_list — Function

Get the list of lower bounds.

Boscia.get_tol — Method

Get solving tolerance for the BLMO.

Boscia.get_upper_bound_list — Function

Get the list of upper bounds.

Boscia.get_variables_pointers — Method

List of all variable pointers. Depends on how you save your variables internally. In the easy case, this is simply collect(1:N).

Is used in find_best_solution.

Boscia.has_integer_constraint — Function

Has variable an integer constraint?

Boscia.indicator_present — Method

Are indicator constraints present?

Boscia.is_bound_in — Function

To check if there is bound for the variable in the global or node bounds.

Boscia.is_constraint_on_int_var — Function

Check if the subject of the bound cidx is an integer variable (recorded in intvars).

Boscia.is_indicator_feasible — Method

Is a given point v indicator feasible, i.e. meets the indicator constraints? If applicable.

Boscia.is_linear_feasible — Function

Is a given point v linear feasible for the model? That means does v satisfy all bounds and other linear constraints?

Boscia.is_valid_split — Method

Check whether a split is valid, i.e. the upper and lower on variable vidx are not the same.

Boscia.postsolve — Method

postsolve(tree, result, time_ref, verbose)

Runs the post solve both for a cleaner solutiona and to optimize for the continuous variables if present. Prints solution statistics if verbose is true.

Boscia.set_bound! — Function

Change the value of the bound c_idx.

FrankWolfe.compute_extreme_point — Function

Implement FrankWolfe.compute_extreme_point

Given a direction d solves the problem min_x d^T x where x has to be an integer feasible point