The Bounded Linear Minimization Oracle (BLMO)

This section describes the general interface to the Bounded Linear Minimization Oracle and the preimplemented types of BLMO.

General BLMO interface

Time Tracking Wrapper

Boscia.TimeTrackingLMO — Type

TimeTrackingLMO  <: FW.LMO

An LMO wrapping another one tracking the time, number of nodes and number of calls.

Build LMO

Boscia.build_LMO — Method

Build node LMO from global LMO

Four action can be taken: KEEP - constraint is as saved in the global bounds CHANGE - lower/upper bound is changed to the node specific one DELETE - custom bound from the previous node that is invalid at current node and has to be deleted ADD - bound has to be added for this node because it does not exist in the global bounds (e.g. variable bound is a half open interval globally)

MOI Oracle

Boscia.MathOptBLMO — Type

BoundedLinearMinimizationOracle for solvers supporting MathOptInterface.

Boscia.MathOptBLMO — Method

Build MathOptBLMO from FrankWolfe.MathOptLMO.

Base.convert — Method

Convert object of Type MathOptLMO into MathOptBLMO and viceversa.

Bonobo.get_branching_variable — Method

Behavior for strong branching.

Boscia.add_bound_constraint! — Method

Add bound constraint.

Boscia.build_LMO_correct — Method

Check if the bounds were set correctly in build_LMO. Safety check only.

Boscia.build_global_bounds — Method

Read global bounds from the problem

Boscia.check_feasibility — Method

Check if problem is bounded and feasible, i.e. no contradicting constraints.

Boscia.check_infeasible_vertex — Method

Deal with infeasible vertex if necessary, e.g. check what caused it etc.

Boscia.delete_bounds! — Method

Delete bounds.

Boscia.explicit_bounds_binary_var — Method

Add explicit bounds for binary variables.

Boscia.find_best_solution — Method

Find best solution from the solving process.

Boscia.find_best_solution — Method

Finds the best solution in the Optimizer's solution storage, based on the objective function f. Returns the solution vector and the corresponding best value.

Boscia.free_model — Method

Free model data from previous solve (if necessary).

Boscia.get_BLMO_solve_data — Method

Get solve time, number of nodes and number of simplex iterations.

Boscia.get_binary_variables — Method

Get list of binary and integer variables, respectively.

Boscia.get_bound — Method

Read bound value for c_idx.

Boscia.get_int_var — Method

Get the index of the integer variable the bound is working on.

Boscia.get_list_of_variables — Method

Get list of variables indices and the total number of variables. If the problem has n variables, they are expected to contiguous and ordered from 1 to n.

Boscia.get_lower_bound_list — Method

Get the list of lower bounds.

Boscia.get_tol — Method

Get solving tolerance for the BLMO.

Boscia.get_upper_bound_list — Method

Get the list of upper bounds.

Boscia.get_variables_pointers — Method

List of all variable pointers. Depends on how you save your variables internally.

Is used in find_best_solution.

Boscia.has_binary_constraint — Method

Has variable a binary constraint?

Boscia.has_integer_constraint — Method

Does the variable have an integer constraint?

Boscia.indicator_present — Method

Are indicator constraints present?

Boscia.is_bound_in — Method

To check if there is bound for the variable in the global or node bounds.

Boscia.is_constraint_on_int_var — Method

Check if the subject of the bound cidx is an integer variable (recorded in intvars).

Boscia.is_indicator_feasible — Method

Is a given point v indicator feasible, i.e. meets the indicator constraints? If applicable.

Boscia.is_linear_feasible — Method

Is a given point v linear feasible for the model?

Boscia.is_valid_split — Method

Check whether a split is valid, i.e. the upper and lower on variable vidx are not the same.

Boscia.set_bound! — Method

Change the value of the bound c_idx.

Boscia.solve — Method

Solve function in case of MathOptLMO. Converts the lmo into a MathOptBLMO and calls the solve function below.

FrankWolfe.compute_extreme_point — Method

compute_extreme_point

Is implemented in the FrankWolfe package in file "moi_oracle.jl".

Managed BLMO

Boscia.ManagedBoundedLMO — Type

ManagedBoundedLinearMinimizationOracle

A Bounded Linear Minimization Oracle that manages the bounds.

simplelmo - An LMO of type Simple Boundable LMO (see above). lowerbounds - List of lower bounds for the integer variables recorded in intvars. If there is no specific lower bound, set corresponding entry to -Inf. upperbounds - List of upper bounds for the integer variables recorded in intvars. If there is no specific upper bound, set corresponding entry to Inf. n - Total number of variables. intvars - List of indices of the integer variables. solvingtime - The time evaluate `computeextreme_point`.

Boscia.SimpleBoundableLMO — Type

SimpleBoundableLinearMinimizationOracle

A simple LMO that computes the extreme point given the node specific bounds on the integer variables. Can be stateless since all of the bound management is done by the ManagedBoundedLMO.

Boscia.bounded_compute_extreme_point — Function

Computes the extreme point given an direction d, the current lower and upper bounds on the integer variables, and the set of integer variables.

Boscia.is_simple_linear_feasible — Function

Checks whether a given point v is satisfying the constraints on the problem. Note that the bounds on the integer variables are being checked by the ManagedBoundedLMO and do not have to be check here.

Polytopes

Boscia.CubeSimpleBLMO — Type

CubeSimpleBLMO{T}(lower_bounds, upper_bounds)

Hypercube with lower and upper bounds implementing the SimpleBoundableLMO interface.

Boscia.ProbabilitySimplexSimpleBLMO — Type

ProbablitySimplexSimpleBLMO(N)

Scaled Probability Simplex: ∑ x = N.

Boscia.UnitSimplexSimpleBLMO — Type

UnitSimplexSimpleBLMO(N)

Scaled Unit Simplex: ∑ x ≤ N.

Boscia.bounded_compute_extreme_point — Method

If the entry is positve, choose the lower bound. Else, choose the upper bound.

Boscia.bounded_compute_extreme_point — Method

Assign the largest possible values to the entries corresponding to the smallest entries of d.

Boscia.bounded_compute_extreme_point — Method

For all positive entries of d, assign the corresponding lower bound. For non-positive entries, assign largest possible value in increasing order.

Boscia.rounding_hyperplane_heuristic — Method

Hyperplane-aware rounding for the probability simplex.

Boscia.rounding_hyperplane_heuristic — Method

Hyperplane-aware rounding for the unit simplex.