Algorithm Interface

Boscia's solve function only requires the oracles of the objective function f and its gradient g as well as the BLMO encoding the feasible region.

Boscia.postsolve — Method

postsolve(tree, result, time_ref, verbose, max_iteration_post)

Runs the post solve to optimize for the continuous variables if present. Is called if use_post_solve is enabled in the solve function. Prints solution statistics if verbose is set to true.

Boscia.solve — Method

solve(f, g, blmo::BoundedLinearMinimizationOracle; ...)

Requires

f oracle of the objective function.

g oracle of the gradient of the objective

blmo encodes the feasible region and can handle additional bound constraints. This can either be a MIP solver instance (e.g., SCIP) or be a custom type (see polytope_blmos.jl). Has to be of type BoundedLinearMinimizationOracle (see blmo_interface.jl).

Returns

x the best solution found.

tlmo the BLMO wrapped in a TimeTrackingLMO instance.

result a dictionary containg the statistics like number of nodes, total solving etc. It also contains information for plotting progress plots like the lower and upper bound progress.

Optional settings

settings_bnb dictionary of settings for the branch-and-bound algorithm. Created via settings_bnb().

settings_frank_wolfe dictionary of settings for the Frank-Wolfe algorithm. Created via settings_frank_wolfe().

settings_tolerances dictionary of settings for the tolerances. Created via settings_tolerances().

settings_postprocessing dictionary of settings for the postprocessing. Created via settings_postprocessing().

settings_heuristic dictionary of settings for the heuristics. Created via settings_heuristic().

settings_tightening dictionary of settings for the tightening. Created via settings_tightening().

settings_domain dictionary of settings for the domain. Created via settings_domain().