Branch and Bound Tree functionality

This section contains the customization of the branch-and-bound implementation of Bonobo.jl.

Custom Bonobo

Bonobo.optimize! — Method

optimize!(tree::BnBTree; callback=(args...; kwargs...)->())

Optimize the problem using a branch and bound approach. The steps, repeated until terminated is true, are the following:

# 1. get the next open node depending on the traverse strategy
node = get_next_node(tree, tree.options.traverse_strategy)
# 2. evaluate the current node and return the lower and upper bound
# if the problem is infeasible both values should be set to NaN
lb, ub = evaluate_node!(tree, node)
# 3. update the upper and lower bound of the node struct
set_node_bound!(tree.sense, node, lb, ub)
# 4. update the best solution
updated = update_best_solution!(tree, node)
updated && bound!(tree, node.id)
# 5. remove the current node
close_node!(tree, node)
# 6. compute the node children and adds them to the tree
# internally calls get_branching_variable and branch_on_variable!
branch!(tree, node)

A callback function can be provided which will be called whenever a node is closed. It always has the arguments tree and node and is called after the node is closed. Additionally the callback function must accept additional keyword arguments (kwargs) which are set in the following ways:

If the node is infeasible the kwarg node_infeasible is set to true.

If the node has a higher lower bound than the incumbent the kwarg worse_than_incumbent is set to true.

Integer Bounds

Boscia.IntegerBounds — Type

IntegerBounds

Keeps track of the bounds of the integer (binary) variables.

lower_bounds dictionary of Float64, index is the key. upper_bounds dictionary of Float64, index is the key.

Problem

Boscia.SimpleOptimizationProblem — Type

Represents an optimization problem of the form:

min_x f(x)
s.t.  x ∈ X (given by the LMO)
      x_j ∈ Z ∀ j in integer_variables

Boscia.Solve_Stage — Type

Enum for the solving stage

Bonobo.get_branching_indices — Method

Returns the indices of the discrete variables for the branching in Bonobo.BnBTree

Boscia._trivial_domain — Method

Trivial domain function.

Boscia._trivial_domain_point — Method

Trivial domain point function.

Boscia.indicator_present — Method

Are indicator constraints present

Boscia.is_integer_feasible — Method

Checks if a given vector is valid integral solution. Specifically for mixed problems.

Boscia.is_linear_feasible — Method

Checks if x is valid for all linear and variable bound constraints